pyproject.toml
[project]
name = "vlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "moviepy>=2.2.1",
    "openai>=2.8.1",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.2.1",
]
